Moved Exp to better spot, Removed unnecesary files
