+++ src/SuperGamblino.Commands/Commands/SlotsCommand.cs
@@ -4,7 +4,6 @@ using SuperGamblino.Core.Entities;
 using SuperGamblino.Core.GamesObjects;
 using SuperGamblino.Core.Helpers;
 using SuperGamblino.Infrastructure.Connectors;
-using SuperGamblino.Infrastructure.DatabaseObjects;
 using SuperGamblino.Messages;
 
 namespace SuperGamblino.Commands.Commands
@@ -57,8 +56,9 @@ namespace SuperGamblino.Commands.Commands
                         await _usersConnector.GiveCredits(userId, pointsResult + bet);
                     }
 
-                    var expHelper = new Exp(_usersConnector);
-                    var expResult = await expHelper.Give(userId, bet);
+                    var exp = ExpHelpers.CalculateBet((await _usersConnector.GetUser(userId)).Level, bet);
+                    var expResult = await _usersConnector.CommandGiveUserExp(userId, exp);
+                    
                     await _gameHistoryConnector.AddGameHistory(new GameHistory
                     {
                         GameName = "slots",
