--- src/SuperGamblino.Infrastructure/DatabaseConnectosHelper.cs
@@ -1,21 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using SuperGamblino.Infrastructure.Connectors;
-
-namespace SuperGamblino.Infrastructure
-{
-    public static class DatabaseConnectorsHelper
-    {
-        /// <summary>
-        ///     Should be called AFTER ConnectionString service and Logger service
-        /// </summary>
-        /// <param name="serviceCollection"></param>
-        /// <returns></returns>
-        public static IServiceCollection AddDatabaseConnectors(this IServiceCollection serviceCollection)
-        {
-            return serviceCollection
-                .AddTransient<GameHistoryConnector>()
-                .AddTransient<CoindropConnector>()
-                .AddTransient<UsersConnector>();
-        }
-    }
-}
