--- src/SuperGamblino.Infrastructure/DatabaseObjects/DateTimeResult.cs
@@ -1,16 +1,0 @@
-﻿﻿using System;
-
- namespace SuperGamblino.Infrastructure.DatabaseObjects
-{
-    public class DateTimeResult
-    {
-        public DateTimeResult(bool successful, DateTime? dateTime)
-        {
-            Successful = successful;
-            DateTime = dateTime;
-        }
-
-        public bool Successful { get; set; }
-        public DateTime? DateTime { get; set; }
-    }
-}
