+++ tests/SuperGamblino.Commands.Tests/VoteRewardTests.cs
@@ -10,7 +10,6 @@ using SuperGamblino.Commands.Commands;
 using SuperGamblino.Core.Configuration;
 using SuperGamblino.Core.Entities;
 using SuperGamblino.Infrastructure.Connectors;
-using SuperGamblino.Infrastructure.DatabaseObjects;
 using Xunit;
 
 namespace SuperGamblino.Commands.Tests
