+++ tests/SuperGamblino.Commands.Tests/WorkRewardTests.cs
@@ -3,7 +3,6 @@ using Moq;
 using SuperGamblino.Commands.Commands;
 using SuperGamblino.Core.Entities;
 using SuperGamblino.Infrastructure.Connectors;
-using SuperGamblino.Infrastructure.DatabaseObjects;
 using Xunit;
 
 namespace SuperGamblino.Commands.Tests
